Arithmetic Evaluator
Here's the first official draft of the arithmetic_evaluator.py file. It
runs in my testing, in the whole it will require the input from the
other files. It can't do long forms, but it can do shorter ones like
(2+3) * (3/5). I'll keep working on it, I just wanted to give a working
update for the rest of the team to work off of.